move pause test to integration test
